requirements.txt
anthropic
fastapi
uvicorn[standard]
playwright
beautifulsoup4
requests
feedparser
apscheduler
python-dotenv
lxml
